Add tests for case where candidate arrives before SLD(answer)
The bug it's checking for manifests as random timeouts on WPT tests
that use RTCPeerConnection-helper's functions to set up connection.

The bug was fixed in another CL, but these tests exercise interesting
functionality, so worth adding.

The purpose of the helper-test is to ensure that if we introduce more
instabilities by changes to the helper, this test should fail first.

Bug: chromium:1122106
Change-Id: Ic59f76fd36f645cee0e08d6dcd69db0c62cbc903